New pages

+++ Src/Core/Chatter.Application/DataTransferObjects/Chats/CreateChatModel.cs
@@ -3,7 +3,7 @@
     public class CreateChatModel
     {
         public string Name { get; set; }
-        public int ChatTypeId { get; set; }
-        public int CreatorId { get; set; }
+        public long ChatTypeId { get; set; }
+        public long CreatorId { get; set; }
     }
 }
